Use long long instead of long
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@132463 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 9a9d6ad61365bb8d98298747034bec42bd1b763a

+++ Rec/Rich/RichRecQC/src/RichPIDQC.h
@@ -225,7 +225,7 @@ namespace Rich
           bool hasR1Gas; ///< Has Rich1 Gas
           bool hasR2Gas; ///< Has Rich1 Gas
         };
-        typedef Rich::Map<Radiators,unsigned long> RadCount;
+        typedef Rich::Map<Radiators,unsigned long long> RadCount;
         /// Tally of tracks which each type of radiator
         RadCount m_radCount;
